Reimplement sparse_reshape middle transform
Avoid deprecated API.
