+++ model-optimizer/extensions/ops/sparse_reshape.py
@@ -42,6 +42,7 @@ class SparseReshape(Op):
     @staticmethod
     def infer(node: Node):
         input_indices_shape = node.in_port(0).data.get_shape()
+        input_indices_value = node.in_port(0).data.get_value()
         input_shape_value = node.in_port(1).data.get_value()
         new_shape_value = node.in_port(2).data.get_value()
         new_shape_shape = node.in_port(2).data.get_shape()
@@ -60,4 +61,6 @@ class SparseReshape(Op):
         output_indices_shape = np.concatenate((input_indices_shape[0:1], new_shape_shape))
         node.out_port(0).data.set_shape(output_indices_shape)
 
-        # TODO: implement for constant input indices value
+        # TODO: implement constant value propogation for common case
+        if np.array_equal(input_shape_value, output_shape_value) and input_indices_value is not None:
+            node.out_port(0).data.set_value(input_indices_value)
